Update README.md
Until recently, automobile manufacturers and dealers have mainly focused on their new vehicle business with the exclusion of used cars, often viewed used cars as a byproduct. However, the competition in the market and the threat of new entries have created a great extent of an upsurge in the used car dealership. Moreover, the added quality and reliability of used cars changed the consumer attitude and increased the sales of the used passenger cars. Investing in used car management has become one of the market's requirements characterized by slimming margin, relentless competition, and demanding consumers.

So, here I used this following dataset to predict the price of any used car depending on there features. We’ll use a dataset which was provided by the Board Infinity comprised of 15157 rows with 8 columns.

The dataset is broken into 11902 records for training and 2976 for testing. The goal is to build a system that can accurately predict the price of car. The competition is evaluated using mse, R2, Adjusted R2. Further, I built a flask API and deployed the model in the Heroku environment.